% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{find_loc}
\alias{find_loc}
\title{Uses googlemap geocoding API to find the lat/lon of the location that would analyze whether for
Sets as default in your environmnet for all other functions in this package
required a google API Key.
Keys have a free trial period and can be acquired at https://developers.google.com/maps/documentation/geocoding/get-api-key
Use the ggmap register_google(key) function to register your key  with your R session}
\usage{
find_loc(location)
}
\arguments{
\item{location}{The location you would like to set the lat/lon for all other functions passed as string. Be as specific as possible}
}
\description{
Uses googlemap geocoding API to find the lat/lon of the location that would analyze whether for
Sets as default in your environmnet for all other functions in this package
required a google API Key.
Keys have a free trial period and can be acquired at https://developers.google.com/maps/documentation/geocoding/get-api-key
Use the ggmap register_google(key) function to register your key  with your R session
}
