% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prev_weather.R
\name{get_celsius}
\alias{get_celsius}
\title{Converts the degrees from Fahrenheit to Celsius}
\usage{
get_celsius(fah)
}
\arguments{
\item{fah}{The temperature in degrees Fahrenheit}
}
\value{
A double of the temperature in degrees celsius
}
\description{
Converts the degrees from Fahrenheit to Celsius
}
