% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{plot_weather}
\alias{plot_weather}
\title{Uses data pulled from the API to generate a plot of the weather for the past number of user inputted days}
\usage{
plot_weather(latitude, longitude, num_days, CF, api_key)
}
\arguments{
\item{latitude}{x}

\item{longitude}{x}

\item{num_days}{x}

\item{CF}{x}

\item{api_key}{x}
}
\value{
plot of temp from the past days
}
\description{
Uses data pulled from the API to generate a plot of the weather for the past number of user inputted days
}
