% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{radius}
\alias{radius}
\title{creates dataframe for use for the plot weather function}
\usage{
radius(latitude, longitude, num_days, api_key)
}
\arguments{
\item{latitude}{latitudinal coordinate of the location}

\item{longitude}{longitudinal coordinate of the location}

\item{num_days}{Number of days of past weather to be viewed}

\item{api_key}{climacell.io API Key}
}
\value{
a dataframe
}
\description{
creates dataframe for use for the plot weather function
}
