% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{register_climacell_key}
\alias{register_climacell_key}
\title{Input API key to be used for the other functions in the tempraturr, API key variable to your global environment}
\usage{
register_climacell_key(apikey)
}
\arguments{
\item{apikey}{Copy and paste climacell API key to be used for use in this function}
}
\value{
variable "apikey" is saved to your environment for further usage of this package
}
\description{
Input API key to be used for the other functions in the tempraturr, API key variable to your global environment
}
