% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_Comparison.R
\name{weather_comparison}
\alias{weather_comparison}
\title{Gives a Data Set of Weather Data for Today(The Upcoming 24 hours ) and Tommorow(The Following 24 hours) with the diffrences between the two days included as well .}
\usage{
weather_comparison(latitude = lat, longitude = lon, api_key = apikey)
}
\arguments{
\item{latitude}{The latitude of the location, default stored in environment by find_loc function}

\item{longitude}{The longitude of the location, default stored in environment by find_loc function}

\item{api_key}{The key for the api, default stored in environment by register_climacell_key function}
}
\value{
A data set with weather information between today and tommorow and the diffrences by hour
}
\description{
Gives a Data Set of Weather Data for Today(The Upcoming 24 hours ) and Tommorow(The Following 24 hours) with the diffrences between the two days included as well .
}
