% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prev_weather.R
\name{get_prev_temp}
\alias{get_prev_temp}
\title{Gives the temperatures of a given location for a given number of days}
\usage{
get_prev_temp(latitude, longitude, num_days = 7, api_key)
}
\arguments{
\item{latitude}{The latitude of the location}

\item{longitude}{The longitude of the location}

\item{num_days}{The number of days viewed}

\item{api_key}{The key for the api}
}
\value{
A data frame of temperatures with dates and observation times in UTC(possibly a warning)
}
\description{
Gives the temperatures of a given location for a given number of days
}
