% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{radius}
\alias{radius}
\title{creates dataframe for use for the plot weather function}
\usage{
radius(latitude, longitude, num_days, api_key)
}
\arguments{
\item{latitude}{x}

\item{longitude}{x}

\item{num_days}{x}

\item{api_key}{x}
}
\value{
a dataframe
}
\description{
creates dataframe for use for the plot weather function
}
